Part B-5 Completed
Not sure how far we are supposed to improve, but i did improve, I tested some variations on features, and found that removing both sulfur values and citric acid, and adding in residual sugar, fixed acidity, sulphates, and ph increased the r-squared value, lowered the mean_squared_error value, and the root_mean_squared_error value. However the bias jumped up quite a bit from -38 to 201 and I'm not sure if this is a major issue or not.