Part 3 of Milestone 3 Completed
A little uncertain of why not all the centroids are being graphed but on specific ones, I looked over some stack overflow code to make the plots work but I am a little lost as I was expecting something similar to the total centroids but it's only using about 20-30% of the total and I do not know why it is doing this. Other then that the first graph is working as intended and I put a reference twice to the stackoverflow post I based it off of.